+++ pure_extensions/lib/pure_extensions.dart
@@ -9,6 +9,7 @@ export 'src/dart/collections/set.dart';
 export 'src/dart/color/color_extension.dart';
 export 'src/dart/color/hsl_color.dart';
 export 'src/dart/primitives/date_time.dart';
+export 'src/dart/primitives/duration.dart';
 export 'src/dart/primitives/geo_point.dart';
 export 'src/dart/primitives/numbers.dart';
 export 'src/dart/primitives/readable_file.dart';